Upgrading code to the future

+++ features/support/env.rb
@@ -12,22 +12,21 @@ require 'pry'
 require 'mail'
 require 'nokogiri'
 
-screenshot_dir = "html-results/screenshots"
+screenshot_dir = 'html-results/screenshots'
 FileUtils.mkdir_p screenshot_dir
 Watir.default_timeout = 20
 
-VW_USER = "vw_envs"
-VW_PASSWORD = "sooT5wioAhxuiH0o"
-
+VW_USER = 'vw_envs'
+VW_PASSWORD = 'sooT5wioAhxuiH0o'
 
 URLS = "http://#{ENV['HOST']}"
 
 DEFAULT_ACCOUNT = {
-    :title => "Mr",
-    :first_name => "AutomatedTestUser",
-    :surname => "AutomatedTestUser",
-    :email_base => "@example.com",
-    :password => "Abcd!2345",
+  title: 'Mr',
+  first_name: 'AutomatedTestUser',
+  surname: 'AutomatedTestUser',
+  email_base: '@example.com',
+  password: 'Abcd!2345'
 }
 
 ffprofile = Selenium::WebDriver::Firefox::Profile.new
@@ -49,7 +48,6 @@ ffprofile['content.notify.backoffcount'] = 5
 ffprofile['gfx.direct2d.disabled'] = true
 ffprofile['layers.acceleration.disabled'] = true
 
-
 require_all 'lib'
 
 if ENV['HEADLESS'] =~ /true/i
@@ -59,18 +57,18 @@ if ENV['HEADLESS'] =~ /true/i
 end
 
 case ENV['BROWSER']
-  when /chrome/i
-    browser = Watir::Browser.start(ENV['HOST'], :chrome)
-    browser.window.resize_to 1024, 768
-  when /ie/i
-    browser = Watir::Browser.new :ie
-    browser.window.resize_to 1024, 768
-  when /firefox/i
-    profile = Selenium::WebDriver::Firefox::Profile.new
-    browser = Watir::Browser.new :firefox, :profile => ffprofile
-  else # Default to Firefox if we don't know exactly what we're using
-    browser = Watir::Browser.new :firefox, :profile => ffprofile
-    browser.window.resize_to 1024, 768
+when /chrome/i
+  browser = Watir::Browser.start(ENV['HOST'], :chrome)
+  browser.window.resize_to 1024, 768
+when /ie/i
+  browser = Watir::Browser.new :ie
+  browser.window.resize_to 1024, 768
+when /firefox/i
+  profile = Selenium::WebDriver::Firefox::Profile.new
+  browser = Watir::Browser.new :firefox, profile: ffprofile
+else # Default to Firefox if we don't know exactly what we're using
+  browser = Watir::Browser.new :firefox, profile: ffprofile
+  browser.window.resize_to 1024, 768
 end
 
 # Auth Handling, forces login to servers in active session
@@ -82,14 +80,13 @@ unless ENV['TAG'].to_sym == :live
 end
 
 class CustomWorld
-  class << self;
+  class << self
     attr_accessor :browser
   end
 
   def site
-    @site ||= (VWBase.new(CustomWorld::browser))
+    @site ||= (VWBase.new(CustomWorld.browser))
   end
-
 end
 
 CustomWorld.browser = browser
@@ -99,26 +96,26 @@ World do
 end
 
 Before('@vw-plugin-cps-vw-section-owners, @vw-web-site, @vw-retailers-site, @my_vw') do
-  browser.execute_script("window.alert = function() {}")
+  browser.execute_script('window.alert = function() {}')
 end
 
 AfterStep('@vw-plugin-cps-vw-section-owners, @vw-web-site, @vw-retailers-site, @my_vw') do
   begin
-    browser.execute_script("window.alert = function() {}")
+    browser.execute_script('window.alert = function() {}')
   rescue Selenium::WebDriver::Error::UnhandledAlertError
-    browser.execute_script("window.alert = function() {}")
+    browser.execute_script('window.alert = function() {}')
     retry
   end
 end
 
 After do |scenario|
   if scenario.failed?
-    time = Time.now.strftime("%Y-%m-%d-%H%M%S")
+    time = Time.now.strftime('%Y-%m-%d-%H%M%S')
     name = "#{time}-#{scenario.name}".slice(0, 250).gsub(/[\,\/]/, '-')
     name.gsub!(/[\|\s\#]/, '')
     file_dir = "#{screenshot_dir}/#{name}.png"
     browser.screenshot.save "#{file_dir}"
-    embed("screenshots/#{name}.png", "image/png" "SCREENSHOT")
+    embed("screenshots/#{name}.png", 'image/png' 'SCREENSHOT')
   end
 end
 
@@ -128,12 +125,10 @@ end
 
 at_exit do
   browser.close
-  if ENV['HEADLESS'] =~ /true/i
-    headless.destroy
-  end
+  headless.destroy if ENV['HEADLESS'] =~ /true/i
   begin
     require_relative '../../HTMLReport.rb'
   rescue
-    puts "Unable to generate HTML Report - Skipping Report Generation"
+    puts 'Unable to generate HTML Report - Skipping Report Generation'
   end
 end
